seperated callback functions to controllers backen
